removed cast

+++ src/realm/group_writer.cpp
@@ -487,7 +487,7 @@ size_t GroupWriter::get_free_space() {
         size_t sum = 0;
         for (size_t j = 0; j < m_free_lengths.size(); ++j) {
             int64_t l = to_size_t(m_free_lengths.get(j));
-            sum += static_cast<size_t>(l);
+            sum += l;
         }
         return sum;
     } else {
